To make it easier to write code that operates only on the server, only on the local player, and then propagate changes to everyone else, I've split out the character as follows.
When a player connects, they create:
CharacterController - local only
CharacterGraphics - network, they own it
CharacterLegacy - network, they own it (this contains everything in the "old" character model, so we don't lose functionality)

When a player connects, the server creates:
CharacterServer - local to the server only
CharacterNetwork - network, the server owns it

This means that clients see this for other players:
CharacterGraphics
CharacterLegacy
CharacterNetwork

So if you wish to write code that operates only on the network for a character, attach it to CharacterServer.
If you wish to push code from this down to the clients from the server, use CharacterNetwork.
If you wish to adjust player controls, use CharacterController.

I have also added some useful utility functions, for doing things like searching for a component in a GameObject's hierarchy.